Real bigint & numeric with exact precision (string-backed)
- bigint (int8) is BigInt-backed: 64-bit values keep full precision;
  large integer literals are bigint (via parser valueText)
- numeric/decimal is arbitrary-precision (Decimal): scale rounding
  (half away from zero), division to 20 digits, numeric(p,s) config
- both return as strings by default, like node-postgres
- type-aware arithmetic operators (int/float stay JS numbers; bigint/
  numeric compute exactly), int4 overflow raises 'integer out of range'
- integer/float behaviour unchanged; sum/count stay number-typed to
  preserve existing semantics

numeric-types 6/6, conformance 93.9% -> 98.0% differential vs pg 16.

+++ src/transforms/aggregations/count.ts
@@ -25,7 +25,7 @@ class CountStar implements AggregationComputer<number> {
     }
 
     get type(): _IType<any> {
-        return Types.bigint;
+        return Types.integer;
     }
 
     computeFromIndex(key: IndexKey, index: _IIndex, t: _Transaction) {
@@ -53,7 +53,7 @@ class CountExpr implements AggregationComputer<number> {
     }
 
     get type(): _IType<any> {
-        return Types.bigint;
+        return Types.integer;
     }
 
     createGroup(t: _Transaction): AggregationGroupComputer<number> {
